+++ api/views/__init__.py
@@ -1,3 +1,3 @@
 from .account import *  # noqa
 from .message import *  # noqa
-from .fake_data import *  # noqa
+from .sample_data import *  # noqa
